Pulling in supermarket version changes
This commit brings in the code from the latest version deployed on
supermarket.chef.io with the latest github repository I was able to
find a fork (dyninc/daemontools). I'm sure there are missing history,
changes, and commits, but this is pretty good for a repository and
cookbook that hasn't changed in six years, I suppose.

Signed-off-by: Joshua Timberman <joshua@chef.io>